requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chromedriver==2.24.1
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.54.1
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.1.0
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.0
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tornado==6.4.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.17.1
zipp==3.20.2